Remove stack nav for explore & notifications; rename main stack nav
