+++ src/navigation/rootNavigation.js
@@ -22,10 +22,8 @@ import ObsEditProvider from "../providers/ObsEditProvider";
 import PhotoGalleryProvider from "../providers/PhotoGalleryProvider";
 import colors from "../styles/colors";
 import { viewStyles } from "../styles/navigation/rootNavigation";
-import ExploreStackNavigator from "./exploreStackNavigation";
 import IdentifyStackNavigation from "./identifyStackNavigation";
-import NotificationsStackNavigation from "./notificationsStackNavigation";
-import ObservationsStackNavigation from "./observationsStackNavigation";
+import MainStackNavigation from "./mainStackNavigation";
 import ProjectsStackNavigation from "./projectsStackNavigation";
 
 // this removes the default hamburger menu from header
@@ -93,39 +91,28 @@ const App = ( ): React.Node => {
                   drawerContent={drawerRenderer}
                 >
                   <Drawer.Screen
-                    name="observations"
-                    component={ObservationsStackNavigation}
+                    name="MainStack"
+                    component={MainStackNavigation}
                     options={hideHeader}
                   />
-                  <Drawer.Screen
-                    name="notifications"
-                    component={NotificationsStackNavigation}
-                    options={hideHeader}
-                  />
+                  <Drawer.Screen name="search" component={Search} />
                   <Drawer.Screen
                     name="identify"
                     component={IdentifyStackNavigation}
                     options={hideHeader}
                   />
-                  <Drawer.Screen name="search" component={Search} />
                   <Drawer.Screen
                     name="projects"
                     component={ProjectsStackNavigation}
                     options={hideHeader}
                   />
                   <Drawer.Screen name="settings" component={Settings} options={hideHeader} />
-                  <Drawer.Screen name="following (dashboard)" component={PlaceholderComponent} />
                   <Drawer.Screen
                     name="about"
                     component={About}
                   />
-                  <Drawer.Screen name="help/tutorials" component={PlaceholderComponent} />
+                  <Drawer.Screen name="help" component={PlaceholderComponent} />
                   <Drawer.Screen name="login" component={MortalLogin} options={hideHeader} />
-                  <Drawer.Screen
-                    name="explore stack"
-                    component={ExploreStackNavigator}
-                    options={hideHeader}
-                  />
                   <Drawer.Screen
                     name="network"
                     component={NetworkLogging}
